Move select components into directory
